feat!: Improvements for MenuItem and AppMenu.
BREAKING CHANGE: In MenuItem to manage the selection, we use now 'onClick' event instead of 'onChange' one, because we don't render a hidden input as a child anymore due to accessibility reasons. Now the selections is totally controlled and it's based in 'onClick' event, 'selectionScheme' prop to define if it's single or multiple selection, and 'state' prop which set 'aria-checked' to true if its value is 'selected'.