feat!(datetime): replaced hasLeftHoverEffect & hasRightHoverEffect by selectionLength
The parameter `hasLeftHoverEffect` & `hasRightHoverEffect` of Calendar
component that usually is returned by the hooks `useCalendarSingle` (as
`false`) and the `useCalendarRange` (as `true`) are replaced by the
`selectionLength` prop that in the single case should be passed as `1` and in
the range case should be passed as `2`. This new parameter is not provided by
the hooks anymore, and by default it is `1`.